--- Game Jam Winner/Assets/SceneSwap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aacb49028b3e2a54bbda94ce4aa14ea8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
